Add files via upload
This is a osmw calculator made with the exact data that the server is currently running on with. 
If any changes will occur in the game the calculator will receive a new version as well in due time.
Fell free to let me know if there's any discrepancies you might notice.

FYI: several pet images have not yet been added eg. Hell Devil.